修改cnn为合理的包名
