+++ funasr/models/campplus/__init__.py
@@ -1,0 +1,1 @@
+from .campplus import CAMPPlus
