--- funasr/models/cnn/__init__.py
@@ -1,3 +1,0 @@
-from .DTDNN import CAMPPlus
-from .ResNet import ERes2Net
-from .ResNet_aug import ERes2NetAug
