+++ funasr/models/eres2net/__init__.py
@@ -1,0 +1,2 @@
+from .eres2net import ERes2Net
+from .eres2net_aug import ERes2NetAug
